refactor(twoLevels): split formPromisesManager
