+++ javascripts/enumlevel2Field/input.js
@@ -7,6 +7,7 @@
  * file that was distributed with this source code.
  */
 import allEntriesLoader from '../allEntriesLoader.service.js'
+import formPromisesManager from '../formPromises.service.js'
 import twoLevelsHelper from '../twolevels.js'
 
 // create a new Event `propChange`
@@ -637,7 +638,7 @@ const enumlevel2Helper = {
             if (typeof parentsFields != "object"){
                 throw "'parentsFields' should be an object with format 'fieldName' => field"
             } else {
-                let promisesData = twoLevelsHelper.initPromisesData()
+                let promisesData = formPromisesManager.initPromisesData()
                 for (let fieldName in parentsFields){
                     let parentField = parentsFields[fieldName]
                     if (parentField && parentField.linkedObjectId.length > 0){
@@ -651,7 +652,7 @@ const enumlevel2Helper = {
                                 )
                             )
                         if (canGetSecondValuesByForm){
-                            twoLevelsHelper.createPromise(promisesData,{
+                            formPromisesManager.createPromise(promisesData,{
                                 formId: parentField.linkedObjectId,
                                 processFormAsync: async (form)=>{
                                     return twoLevelsHelper.getAvailableSecondLevelsValues(form,parentField,values,this.levels2)
@@ -668,7 +669,7 @@ const enumlevel2Helper = {
                         } else {
                             for (let formIdData of formsIds){
                                 const formId = formIdData.id
-                                twoLevelsHelper.createPromise(promisesData,{
+                                formPromisesManager.createPromise(promisesData,{
                                     formId,
                                     processFormAsync: async (form)=>{
                                         return twoLevelsHelper.getAvailableSecondLevelsValuesForLists(
@@ -693,7 +694,7 @@ const enumlevel2Helper = {
                         }
                     }
                 }
-                return await twoLevelsHelper.resolvePromises(promisesData)
+                return await formPromisesManager.resolvePromises(promisesData)
             }
         },
         updateRadio(field,secondLevelValues,childId,parentValuesAssociations){
