+++ javascripts/twolevels.js
@@ -10,7 +10,6 @@
 import eventDispatcher from './eventDispatcher.service.js'
 import allEntriesLoader from './allEntriesLoader.service.js'
 import entryLoader from './entryLoader.service.js'
-import formLoader from './formLoader.service.js'
 
 let correspondances = {}
 let parentEntries = {}
@@ -441,61 +440,10 @@ const getAvailableSecondLevelsValuesForLists = async (associatingForm,fieldName,
     return [secondLevelValues,associatingForm,associations]
 }
 
-// manage promise
-const initPromisesData = () => {
-    return {
-        promises: [],
-        promisesLabel: []
-    }
-}
-
-const createPromise = (promisesData,{formId,processFormAsync,getEntriesAsync,getEntriesLabel}) => {
-    // get form
-    promisesData.promises.push(new Promise((resolve,reject)=>{
-        formLoader.load(formId).then((form)=>{
-            processFormAsync(form).then(([,formModified,])=>{
-                resolve(formModified)
-            })
-            .catch((e)=>reject(e))
-        })
-        .catch((e)=>{
-            reject(e)
-        })
-    }))
-    promisesData.promisesLabel.push(`getting form ${formId}`)
-    // start getting entries in parallel of form
-    promisesData.promises.push(
-        new Promise((resolve,reject)=>{
-            getEntriesAsync().then((entries)=>{
-                resolve(entries)
-            })
-            .catch((e)=>{
-                reject(e)
-            })
-        })
-    )
-    promisesData.promisesLabel.push(getEntriesLabel)
-}
-
-const resolvePromises = async (promisesData) => {
-    return await Promise.allSettled(promisesData.promises).then((promisesStatus)=>{
-        promisesStatus.forEach((p,idx)=>{
-            if (p.status != "fulfilled"){
-                console.warn(`error : ${p.reason} (when ${promisesData.promisesLabel[idx]})`)
-                console.error({error:p.reason})
-            }
-        })
-        return promisesStatus
-    })
-}
-
 export default {
-    createPromise,
     formatChildField,
     formatParentField,
     getAvailableSecondLevelsValues,
     getAvailableSecondLevelsValuesForLists,
-    getParentEntries,
-    initPromisesData,
-    resolvePromises
+    getParentEntries
 }
